Add Andrea Frittoli to the members list
I work on open source for IBM, I'm one of the maintainers of Tekton
and member of the governance board. I implemented the cloud events
and I'm interested make Tekton interoperable with other solution
in the CI/CD space.

I've been contributing to the events in CI/CD subgroup, and I'm
interested in collaborating with the interop SIG in general.