Create placeholder for archiving 2020 meeting minutes
Working on HackMD got slower due to length of the document so
the plan is to move/archive the minutes of meetings held during
2020 into this newly created document in order to help with the
responsiveness of HackMD while working on the document.

The documents will be cross-linked to each other so people will
be able to access the archived minutes as well as the latest ones.

The document on HackMD that will be linked to this file is available
on https://hackmd.io/niM1USFoQyycav8IczCmzQ?view
This document on HackMD will be locked to avoid accidental updates
to it.

The SIG will continue using the original document during the new
year once the minutes from the latest 2020 meeting are committed to
repo: https://hackmd.io/HuufSDMaTPyb3qxkyBKg3A?view