Split vocabulary.md into two documents
This PR aims to start conversation around how the vocabulary.md
document could be split to two for ease of read as well as for
publishing on CDF Best Practices site.

Here is the update
- vocabulary.md: A note is included in the document to point people
to the right document so they can still find what they are looking for.
- tools-terminology.md: Terms used by CI/CD tools are included in this
document and could perhaps be considered to be published on CDF Best
Practices site's technology related pages.
- pipelines-terminology.md: Terms used for common pipeline stages and
steps are included in this document and could perhaps be considered to
be published on corresponding places on CDF Best Practices site.